+++ contact-management/contact-management/Controller/AddContactViewController.swift
@@ -8,10 +8,9 @@
 import UIKit
 
 final class AddContactViewController: UIViewController {
-    private var contactListStorage: ContactListStorage?
-    private var nameContact: String?
-    private var phoneContact: String?
-    private var ageContact: Int?
+    var nameContact: String?
+    var phoneContact: String?
+    var ageContact: Int?
     
     @IBOutlet weak var nameLabel: UILabel!
     @IBOutlet weak var ageLabel: UILabel!
@@ -21,7 +20,8 @@ final class AddContactViewController: UIViewController {
     @IBOutlet weak var ageTextField: UITextField!
     @IBOutlet weak var phoneTextField: UITextField!
     
-    private var phoneFormat: PhoneFormat?
+    private var contactListStorage: ContactListStorage?
+    var phoneFormat: PhoneFormat?
     
     required init?(coder: NSCoder) {
         self.contactListStorage = nil
@@ -37,6 +37,7 @@ final class AddContactViewController: UIViewController {
     
     override func viewDidLoad() {
         setView()
+        addTextField()
     }
     
     private func setView() {
@@ -44,10 +45,6 @@ final class AddContactViewController: UIViewController {
         nameLabel.text = "이름"
         ageLabel.text = "나이"
         phoneLabel.text = "연락처"
-        
-        nameTextField.addTarget(self, action: #selector(nameFieldDidChange(_:)), for: .editingChanged)
-        ageTextField.addTarget(self, action: #selector(ageFieldDidChange(_:)), for: .editingChanged)
-        phoneTextField.addTarget(self, action: #selector(phoneFieldDidChange(_:)), for: .editingChanged)
     }
     
     private func unWrappedSender(name: String?, phone: String?, age: Int?) throws -> ContactList {
@@ -70,9 +67,6 @@ extension AddContactViewController {
     }
     
     @IBAction func didTappedSave(_ sender: Any) {
-        let _ : AlertActionHandler = { [weak self] _ in
-            self?.dismiss(animated: true)
-        }
         do {
             let unWrappedResult = try unWrappedSender(
                 name: nameContact,
@@ -111,22 +105,4 @@ extension AddContactViewController {
     }
 }
 
-extension AddContactViewController {
-    @objc func nameFieldDidChange(_ nameField: UITextField) {
-        nameContact = nameField.text
-    }
-    
-    @objc func ageFieldDidChange(_ ageField: UITextField) {
-        guard let age = ageField.text else {
-            return
-        }
-        ageContact = Int(age)
-    }
-    
-    @objc func phoneFieldDidChange(_ phoneField: UITextField) {
-        guard let phone = phoneField.text else { return }
-        phoneField.text = phoneFormat?.addCharacter(at: phone)
-        phoneContact = phoneField.text
-    }
-}
